requirements.txt
pyTelegramBotAPI
kinopoiskpy
pymongo